chore: move interface to SPI
